fix another int/long type mismatch.  This one was causing pigs to die with an fpe on alpha because fscale wasn't properly initted

+++ usr.bin/systat/pigs.c
@@ -66,7 +66,7 @@ static struct p_times {
 } *pt;
 
 static long stime[CPUSTATES];
-static long    fscale;
+static int    fscale;
 static double  lccpu;
 
 WINDOW *
